add colors to movement
