+++ src/index.html
@@ -27,6 +27,9 @@
                             </ul>
                         </fieldset>
                     </form>
+                    <div class="d-grid">
+                        <button class="btn btn-dark" type="button">Clear all completed</button>
+                    </div>
                 </div>
             </div>
         </div>
